\name{stressR-package}
\alias{stressR-package}
\docType{package}
\title{.}
\description{
  .
}
\details{
\tabular{ll}{
  Package: \tab stressR\cr
  Type: \tab Package\cr
  License: \tab GPL (>= 2)\cr
}
}
\references{
  \url{https://github.com/wrathematics/stressR}
}
\author{
  Drew Schmidt \email{wrathematics AT gmail.com}
}
\keyword{Package}
