% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getCreditMarkets}
\alias{getCreditMarkets}
\title{Get credit markets stress components}
\usage{
getCreditMarkets(s = NULL)
}
\arguments{
\item{s}{the list of class \code{stress} from previous queries, or NULL to perform new query}
}
\value{
A list of class \code{stress}
}
\description{
Downloads FRB financial stress index component data.
}
\details{
Downloads the Cleveland FRB data products for financial stress index components daily time series.  Component values include
\itemize{
\item liquidity spread
\item covered interest spread
\item commercial paper - t-bill spread
\item treasury yield curve spread
\item coporate bond spread
}
}
\examples{
\dontrun{
getCreditMarkets()
}
}
\seealso{
getStressData getEquityMarkets getFundingMarkets getForeignExchangeMarkets getRealEstateMarkets getSecuritizationMarkets
}

