% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getFundingMarkets}
\alias{getFundingMarkets}
\title{Get funding markets stress components}
\usage{
getFundingMarkets(s = NULL)
}
\arguments{
\item{s}{the list of class \code{stress} from previous queries, or NULL to perform new query}
}
\value{
A list of class \code{stress}
}
\description{
Downloads FRB financial stress index component data.
}
\details{
Downloads the Cleveland FRB data products for financial stress index components daily time series.  Component values include
\itemize{
\item financial beta
\item interbank cost of borrowing
\item bank bond spread
\item interbank liquidity spread
}
}
\examples{
\dontrun{
getFundingMarkets()
}
}
\seealso{
getStressData getEquityMarkets getCreditMarkets getForeignExchangeMarkets getRealEstateMarkets getSecuritizationMarkets
}

