% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getRealEstateMarkets}
\alias{getRealEstateMarkets}
\title{Get foreign exchange markets stress components}
\usage{
getRealEstateMarkets(s = NULL)
}
\arguments{
\item{s}{the list of class \code{stress} from previous queries, or NULL to perform new query}
}
\value{
A list of class \code{stress}
}
\description{
Downloads FRB financial stress index component data.
}
\details{
Downloads the Cleveland FRB data products for financial stress index components daily time series.  Component values include
\itemize{
\item commecial real estate spread
\item residential real estate spread
}
}
\examples{
\dontrun{
getRealEstateMarkets()
}
}
\seealso{
getStressData getEquityMarkets getCreditMarkets getFundingMarkets  getForeignExchangeMarkets getSecuritizationMarkets
}

