% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getStressIndex}
\alias{getStressIndex}
\title{Get financial stress index data.}
\usage{
getStressIndex(verbose = FALSE)
}
\arguments{
\item{verbose}{whether to print progress messages, default FALSE}
}
\value{
List of class type \code{cfsi} containing \code{xts} time history object \code{df}, plot colors array \code{colors}, default plot main title \code{main}, and default plot y-axis label \code{ylab}.
}
\description{
Downloads Cleveland FRB financial stress index data.
}
\details{
Transforms the HTML into a data frame, transforms the character date into \code{Date} objects, and then an \code{xts} object.
}
\examples{
\dontrun{
getStressIndex()
}
}
\references{
\url{http://www.clevelandfed.org/research/data/financial_stress_index/index.cfm}
}
\seealso{
getStressComponents
}

