% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{stressR}
\alias{stressR}
\alias{stressR-package}
\title{stressR-package}
\description{
TODO.
}
\details{
TODO.

\tabular{ll}{ Package: \tab stressR\cr Type: \tab Package\cr License: \tab
GPL (>= 2)\cr }
}
\author{
Drew Schmidt \email{wrathematics AT gmail.com}
}
\references{
\url{https://github.com/wrathematics/stressR}
}
\keyword{Package}

