% Generated by roxygen2 (4.0.1): do not edit by hand
\name{xyplot.cfsi}
\alias{xyplot.cfsi}
\title{Financial stress index component data xyplot}
\usage{
\method{xyplot}{cfsi}(x, ...)
}
\arguments{
\item{x}{an object of class \code{stress} as returned by \code{\link[stressr]{getStressIndex}}.}

\item{...}{other parameters passed to \code{\link[lattice]{xyplot}}.}
}
\description{
Provides a convenience function for passing an \code{cfsi} object to \code{xyplot}.
}
\examples{
\dontrun{
ci = getStressIndex()
xyplot(ci)
}
}
\seealso{
stressLineChart stressAreaChart getStressIndex xyplot.stress
}

