% Generated by roxygen2 (4.0.1): do not edit by hand
\name{xyplot.stress}
\alias{xyplot.stress}
\title{Financial stress index component data xyplot}
\usage{
\method{xyplot}{stress}(x, ...)
}
\arguments{
\item{x}{an object of class \code{stress} as returned by \code{\link[stressr]{getStressComponents}} and its many offspring.}

\item{...}{other parameters passed to \code{\link[lattice]{xyplot}}.}
}
\description{
Provides a convenience function for passing an \code{stress} object to \code{xyplot}.
}
\examples{
\dontrun{
require(lattice)
fs <- getFundingStress()
xyplot(fs)
}
}
\seealso{
stressLineChart stressAreaChart getStressComponents xyplot.cfsi
}

